format: collapse T.dynamic into existing symbolic API + fix F841 unused locals
- Remove tilelang/language/symbolics.py; upstream already has def symbolic()
  at line 199 of __init__.py. Just add a def dynamic() next to it that
  delegates to symbolic for comma/space-separated names.
- Add # codespell:ignore vor on the npuir_or alias (false positive on the
  pre-existing identifier name).
- Delete unused-local declarations flagged by ruff F841:
  * lighting_indexer_bwd_kernel.py: q_shared_3d, w_shared, grad_shared_loader,
    d_q_shared, d_q_3d
  * sparse_mla_fwd_kernel.py: DV, v in _ref_torch placeholder
- All three production kernels still pass on NPU after these changes.
